Stricter checks in accessors and improved error messages
Accessors check (within what's reasonable for Perl) that values are the correct
type for the fields.

All error messages mention the fully-qualified name of the field.

Remove the promise of deep checks/coersion for repeated/message setters (too
expensive as a default, and it's probably better implemented as a separate
method anyway).