Add an option to emulate proto3 map API under proto2
It requires the proto2 file to reproduce the message structure produced by the
proto3 compiler when it sees the map<> syntax.

Most encoding/decoding tests have been moved to a common proto2/proto3 test
file because the map<> syntax is syntactic sugar, so the encoding/decoding code
is the same for proto2 and proto3.