Don't croak() from error collector
Because this will unwind the C stack and leave the protobuf descriptor
in an inconsistent state that will trigger an assert when it is
destroyed.

And due to unfortunate ordering, it also means that most of the time the
assert is triggered before Perl has printed out the error message from
croak().