Avoid emitting empty packed repeated fields
They are illegal [1], and decoders break when they (rightfully) expect
to always find a value to decode after the beginning of a packed
repeated field is encountered.

This also affects how repeated fields are encoded to JSON: previously,
an empty repeated field would yield `{"field_name": []}`, while now an
empty repeated field is omitted from the JSON output. This follows the
scheme of the binary encoding, and is not specified by the protocol
buffers JSON encoding description [2]. Usual protocol buffers default
value handling is described, with `null` additionally accepted to
indicate a default value should be used. Given the absense of evidence
to the contrary, omitting empty lists for repeated fields appears to be
legal and in the spirit of the specified JSON encoding.

I'd rather avoid such a change to the JSON encoding, but upb seems
committed to emitting something as soon as `sink->StartSequence` is
called, and it doesn't give us a way to detect if we're writing JSON or
binary encoding anyway.

[1] "A packed repeated field containing zero elements does not appear in
the encoded message.", from
https://developers.google.com/protocol-buffers/docs/encoding#packed_repeated_fields

[2] https://developers.google.com/protocol-buffers/docs/proto3#json_mapping