Add a more efficient, lower level decoder transform function
With the current transform function, the decoder still needs to build a full
Perl data structure for the value as encoded, part of which might be thrown
away by the transformation.

Field tables avoid the need to build the data structure, and provide a
(usually) faster way for the transformation function to look up fields.