Use booleans in service introspection test
With the release of 5.36.0, Perl got real boolean values which
try to behave like the traditional `1` and `''` in older versions.
However, starting with 5.37.2, Data::Dumper stringifies these
differently. "True", for example, goes from `1` to being `!!1`.

This is of no consequence for Perl, but was resulting in the
service introspection test failing in versions of Perl greater than
5.37.2 because the changes in the built-in Data::Dumper were making
Test::Differences report a false positive. See [1] for some additional
context.

This patch fixes this by storing the result of what the current version
of Perl considers to be "true" and "false" in constants, and using
those in the test whenever they are needed. This makes the tests pass
again under newer perls.

[1]: https://github.com/DrHyde/perl-modules-Test-Differences/issues/21